Fix for random crashes when converting portable to windows PDBs
This fixes issue https://github.com/dotnet/buildtools/issues/1737
Tracked it down by using gflags page heap (which makes native memory untouchable after free)
as well as GCStress3 (which causes many GCs.  Then modified Pdb2pdb to convert many PDBs
repeatedy in a loop.  In this confirmation, the app crashes within seconds.   See issue above
for more.

Issue was that SymDocumentWriter's refcount was going to zero while the SymWriter still had
references to it (and would try to close it).   Presumably SymWriter was not correctly AddRefing
its reference.    However since a SymUnmanagedWriterImpl (which is a wrapper around SymWriter
ALREADY had a list of these documents (which will keep the SymDocumentWriters alive), we just
needed to delay clearing them until AFTER the symWriter is dead.    This fixes the bug.

Note when running my modified Pdb2Pdb I was getting a gradual increase in memory, which
suggests that we are leaking memory.  My guess is that we still have COM lifetime issues (leaks)
but that is a less severe issue that we can attack as needed.